Move segment type definitions into a module
Makes them easier to find I think
